Move config.py and run.py into __main__.py
It's not too long; make it one file.
